ajout like/dislike + gestion des erreurs

+++ Controllers/sauces.js
@@ -10,7 +10,7 @@ exports.createSauce = (req, res, next) => {
   });
   sauce.save()
     .then(() => res.status(201).json({ message: 'Objet enregistré !'}))
-    .catch(error => res.status(400).json({ error }));
+    .catch(error => res.status(500).json({ error }));
 };
 
 exports.modifySauce = (req, res, next) => {
@@ -19,19 +19,19 @@ exports.modifySauce = (req, res, next) => {
       ...JSON.parse(req.body.sauce),
       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : { ...req.body };
-  sauces.updateOne({ _id: req.params.id }, { ...saucesObject, _id: req.params.id })
+  sauces.updateOne({ _id: req.params.id, userId:req.auth.userId }, { ...saucesObject, _id: req.params.id })
     .then(() => res.status(200).json({ message: 'Objet modifié !'}))
-    .catch(error => res.status(400).json({ error }));
+    .catch(error => res.status(500).json({ error }));
 };
 
 exports.deleteSauce = (req, res, next) => {
-  sauces.findOne({ _id: req.params.id })
+  sauces.findOne({ _id: req.params.id, userId:req.auth.userId })
     .then(sauces => {
       const filename = sauces.imageUrl.split('/images/')[1];
       fs.unlink(`images/${filename}`, () => {
-        sauces.deleteOne({ _id: req.params.id })
+        sauces.deleteOne({ _id: req.params.id, userId:req.auth.userId })
           .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-          .catch(error => res.status(400).json({ error }));
+          .catch(error => res.status(500).json({ error }));
       });
     })
     .catch(error => res.status(500).json({ error }));
@@ -39,12 +39,15 @@ exports.deleteSauce = (req, res, next) => {
 
 exports.getOneSauce = (req, res, next) => {
     sauces.findOne({ _id: req.params.id })
-      .then(sauces => res.status(200).json(sauces))
-      .catch(error => res.status(404).json({ error }));
+      .then(sauces => {
+        if(!sauces)return res.status(404).json({ error: 'Not Found'})
+        res.status(200).json(sauces)
+      })
+      .catch(error => res.status(500).json({ error }));
 };
 
 exports.getAllSauces = (req, res, next) => {
     sauces.find()
       .then(sauces => res.status(200).json(sauces))
-      .catch(error => res.status(400).json({ error }));
+      .catch(error => res.status(500).json({ error }));
 };
